+++ internal/polkavm/host_call/general_functions.go
@@ -256,7 +256,7 @@ func Lookup(gas polkavm.Gas, regs polkavm.Registers, mem polkavm.Memory, s servi
 	h, o := regs[polkavm.A1], regs[polkavm.A2]
 
 	key := make([]byte, 32)
-	if err := mem.Read(h, key); err != nil {
+	if err := mem.Read(uint32(h), key); err != nil {
 		return gas, regs, mem, polkavm.ErrPanicf(err.Error())
 	}
 
@@ -300,7 +300,7 @@ func Read(gas polkavm.Gas, regs polkavm.Registers, mem polkavm.Memory, s service
 
 	// read key data from memory at ko..ko+kz
 	keyData := make([]byte, kz)
-	err := mem.Read(ko, keyData)
+	err := mem.Read(uint32(ko), keyData)
 	if err != nil {
 		return gas, regs, mem, polkavm.ErrPanicf(err.Error())
 	}
@@ -336,7 +336,7 @@ func Write(gas polkavm.Gas, regs polkavm.Registers, mem polkavm.Memory, s servic
 
 	//µko⋅⋅⋅+kz
 	keyData := make([]byte, kz)
-	err := mem.Read(ko, keyData)
+	err := mem.Read(uint32(ko), keyData)
 	if err != nil {
 		return gas, regs, mem, s, polkavm.ErrPanicf(err.Error())
 	}
@@ -353,7 +353,7 @@ func Write(gas polkavm.Gas, regs polkavm.Registers, mem polkavm.Memory, s servic
 		}
 	} else {
 		valueData := make([]byte, vz)
-		err = mem.Read(vo, valueData)
+		err = mem.Read(uint32(vo), valueData)
 		if err != nil {
 			return gas, regs, mem, s, polkavm.ErrPanicf(err.Error())
 		}
@@ -463,7 +463,7 @@ func Log(gas polkavm.Gas, regs polkavm.Registers, mem polkavm.Memory, core *uint
 	// Write target
 	if to != 0 && tz != 0 {
 		targetBytes := make([]byte, tz)
-		err := mem.Read(to, targetBytes)
+		err := mem.Read(uint32(to), targetBytes)
 		if err != nil {
 			log.VM.Error().Msgf("unable to access memory for target: address %d length %d", to, tz)
 		}
@@ -472,7 +472,7 @@ func Log(gas polkavm.Gas, regs polkavm.Registers, mem polkavm.Memory, core *uint
 	}
 
 	msgBytes := make([]byte, xz)
-	err := mem.Read(xo, msgBytes)
+	err := mem.Read(uint32(xo), msgBytes)
 	if err != nil {
 		log.VM.Error().Msgf("unable to access memory for target: address %d length %d", to, tz)
 	}
